more minor optimizations

+++ src/memory.h
@@ -6,7 +6,7 @@
 // Name: read_byte
 // Desc: reads a value from memory including mapper and I/O
 //------------------------------------------------------------------------------
-inline uint8_t read_byte(uint16_t address) {
+inline uint8_t read_byte(uint32_t address) {
 	return nes::bus::read_memory(address);
 }
 
@@ -22,7 +22,7 @@ inline uint8_t read_byte_zp(uint8_t address) {
 // Name: write_byte
 // Desc: writes a value to memory including mapper and I/O
 //------------------------------------------------------------------------------
-inline void write_byte(uint16_t address, uint8_t value) {
+inline void write_byte(uint32_t address, uint8_t value) {
 	nes::bus::write_memory(address, value);
 }
